refactor: move code from serializers.py for better consistency
