+++ jobMarket/services/analysis/posting_by_job_position/delivery/chart1serializers.py
@@ -1,9 +1,8 @@
 from rest_framework import serializers
 from datetime import datetime
 
-from .models import *
-
-from .constants import JOB_SUB_TYPE_NAMES
+from jobMarket.models import PostingByJobPosition
+from jobMarket.constants import JOB_SUB_TYPE_NAMES
 
 
 class PostingByJobPositionQuerySerializer(serializers.Serializer):
@@ -52,21 +51,3 @@ class PostingByJobPositionSerializer(serializers.ModelSerializer):
     class Meta:
         model = PostingByJobPosition
         fields = ["datetime", "job_type", "count"]
-
-
-class PostingByJobLocationSerializer(serializers.ModelSerializer):
-    location_name = serializers.CharField(
-        label=PostingByJobLocation._meta.get_field("location_name").verbose_name,
-    )
-
-    experience_level = serializers.CharField(
-        label=PostingByJobLocation._meta.get_field("experience_level").verbose_name,
-    )
-
-    rate = serializers.FloatField(
-        label=PostingByJobLocation._meta.get_field("rate").verbose_name,
-    )
-
-    class Meta:
-        model = PostingByJobLocation
-        fields = ["location_name", "experience_level", "rate"]
