+++ jobMarket/services/analysis/posting_by_job_position/delivery/chart1view.py
@@ -2,7 +2,7 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework import status
 from jobMarket.models import PostingByJobPosition
-from jobMarket.serializers import *
+from .chart1serializers import *
 
 from datetime import timedelta
 from django.utils.dateparse import parse_date
